IMPLEMENT CE
The CE algorithm is implemented as well as viz functions to monitor quality.
A trick is used when sampling routes because of a bug in numba (see multinomial).
This trick should not influence optimization. Parralel versions should be considered
for faster algo.
Algo is applied to all benchmark instances and the readme is updated to contain more info.